refactor: split db ebook.rs into focused modules (#202)
Mechanically split the 1,324-line db/src/ebook.rs into a thin module root
(shared types + scan_ebook_library orchestration wrappers + test_support) plus
four submodules under db/src/ebook/: stat (filesystem walk/diff), parse (OPF
metadata + cover-resolution dispatch), cover (sidecar resolution/materialization),
and accent (cover accent-color extraction). Tests distributed next to their
target functions. External paths (omnibus_db::ebook::{scan_ebook_library,
extract_accent, stat_ebook_library, parse_ebook_targets, StatEntry}) preserved
via re-exports. No behavior change.
