+++ db/src/ebook.rs
@@ -16,12 +16,18 @@
 //! permission denied) falls back to the in-memory embedded bytes for the
 //! current scan and retries on the next one.
 
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-use epub::doc::{EpubDoc, EpubVersion};
-use omnibus_shared::{Contributor, EbookMetadata, Identifier};
+use omnibus_shared::EbookMetadata;
 
-use crate::library_layout;
+mod accent;
+mod cover;
+mod parse;
+mod stat;
+
+pub use accent::extract_accent;
+pub use parse::{parse_ebook_targets, ParseTarget};
+pub use stat::{stat_ebook_library, StatEntry, StatScanResult};
 
 /// A single scanner output row — metadata plus the raw cover image bytes
 /// (and mime), if the epub included one. Consumed by [`crate::sync::sync_books`].
@@ -44,56 +50,6 @@ pub struct ScanResult {
     pub path: Option<String>,
     pub books: Vec<IndexedBook>,
     pub error: Option<String>,
-}
-
-/// Phase A output: a single filesystem stat row, with no zip open or OPF
-/// parse. The incremental diff compares these against `book_files` to
-/// classify entries as Unchanged / New / Changed / Removed / Backfill.
-///
-/// An entry with an empty `uuid` is a synthetic placeholder for an
-/// unreadable subdirectory — `error` carries the underlying io message
-/// (e.g. "permission denied" vs. "no such file") so the legacy
-/// [`scan_ebook_library_with`] wrapper can surface it verbatim. The
-/// incremental diff ignores empty-uuid entries.
-#[derive(Debug, Clone, PartialEq)]
-pub struct StatEntry {
-    /// Path relative to the library root — same shape used everywhere else
-    /// as the per-book "filename" string.
-    pub filename: String,
-    /// Stable UUIDv5 of `(library_path, filename)`. Computed here so the
-    /// diff branch never needs to thread `library_path` around again.
-    /// Empty string for placeholder rows (see struct doc).
-    pub uuid: String,
-    /// Filesystem mtime in seconds since the unix epoch, or `0` if
-    /// `entry.metadata().modified()` failed.
-    pub mtime_epoch: i64,
-    /// Filesystem byte size, or `0` if stat failed.
-    pub size_bytes: i64,
-    /// Only populated for placeholder rows — the original io::Error
-    /// message string from the failed `read_dir`. `None` for real epub
-    /// entries.
-    pub error: Option<String>,
-}
-
-/// Phase A result. Mirrors `ScanResult` for the no-path / unreadable-root
-/// failure modes.
-pub struct StatScanResult {
-    pub path: Option<String>,
-    pub entries: Vec<StatEntry>,
-    pub error: Option<String>,
-}
-
-/// Phase B input: one entry per file the diff says needs a full OPF parse
-/// (the New + Changed buckets). The absolute path lets the parser open the
-/// file directly without re-walking; the stat values carry forward into
-/// the resulting `IndexedBook` so the writer persists the same values the
-/// diff observed.
-#[derive(Debug, Clone)]
-pub struct ParseTarget {
-    pub filename: String,
-    pub absolute: PathBuf,
-    pub mtime_epoch: i64,
-    pub size_bytes: i64,
 }
 
 /// Knobs that change how a scan touches the filesystem. Default keeps the
@@ -181,562 +137,15 @@ pub fn scan_ebook_library_with(path: Option<&str>, opts: ScanOptions) -> ScanRes
     }
 }
 
-/// Phase A: walk the library and `stat` every `.epub` without opening the
-/// zip. Returns one `StatEntry` per file plus a synthetic entry (empty
-/// `uuid`) for any unreadable subdirectory — the legacy
-/// `scan_ebook_library_with` shape surfaces these as error rows.
-///
-/// `library_path_key` is what `stable_uuid` hashes; callers pass the same
-/// string they'd use as the `books.library_path` so the uuids line up with
-/// any rows already in the DB.
-pub fn stat_ebook_library(path: Option<&str>, library_path_key: &str) -> StatScanResult {
-    let Some(path_str) = path else {
-        return StatScanResult {
-            path: None,
-            entries: vec![],
-            error: None,
-        };
-    };
-
-    let dir = Path::new(path_str);
-    if !dir.exists() {
-        return StatScanResult {
-            path: Some(path_str.to_string()),
-            entries: vec![],
-            error: Some(format!("path not found: {path_str}")),
-        };
-    }
-
-    let mut entries: Vec<StatEntry> = Vec::new();
-    let mut stack: Vec<PathBuf> = vec![dir.to_path_buf()];
-    while let Some(current) = stack.pop() {
-        let read = match std::fs::read_dir(&current) {
-            Ok(e) => e,
-            Err(e) => {
-                if current == dir {
-                    // Root unreadable: fatal — match legacy behavior.
-                    return StatScanResult {
-                        path: Some(path_str.to_string()),
-                        entries: vec![],
-                        error: Some(format!("could not read directory: {e}")),
-                    };
-                }
-                // Sub-dir unreadable: record a placeholder with an empty
-                // uuid so the legacy wrapper can lift it into an error
-                // row. Carry the io::Error string so callers can
-                // distinguish "permission denied" from "no such file" —
-                // the incremental indexer ignores empty-uuid entries.
-                let relative = current
-                    .strip_prefix(dir)
-                    .unwrap_or(&current)
-                    .to_string_lossy()
-                    .to_string();
-                entries.push(StatEntry {
-                    filename: relative,
-                    uuid: String::new(),
-                    mtime_epoch: 0,
-                    size_bytes: 0,
-                    error: Some(e.to_string()),
-                });
-                continue;
-            }
-        };
-        for entry in read.flatten() {
-            let Ok(file_type) = entry.file_type() else {
-                continue;
-            };
-            let entry_path = entry.path();
-            if file_type.is_dir() {
-                stack.push(entry_path);
-                continue;
-            }
-            if !file_type.is_file() {
-                continue;
-            }
-            let is_epub = entry_path
-                .extension()
-                .and_then(|s| s.to_str())
-                .map(|s| s.eq_ignore_ascii_case("epub"))
-                .unwrap_or(false);
-            if !is_epub {
-                continue;
-            }
-            let relative = entry_path
-                .strip_prefix(dir)
-                .unwrap_or(&entry_path)
-                .to_string_lossy()
-                .to_string();
-            let (mtime_epoch, size_bytes) = stat_file(&entry_path);
-            let uuid = crate::helpers::stable_uuid(library_path_key, &relative);
-            entries.push(StatEntry {
-                filename: relative,
-                uuid,
-                mtime_epoch,
-                size_bytes,
-                error: None,
-            });
-        }
-    }
-
-    entries.sort_by(|a, b| a.filename.cmp(&b.filename));
-
-    StatScanResult {
-        path: Some(path_str.to_string()),
-        entries,
-        error: None,
-    }
-}
-
-/// Phase B: parse the full OPF + cover for the subset of files the diff
-/// said are new or changed. Each target carries the absolute path so we
-/// don't re-walk, and the Phase-A stat values so the resulting
-/// `IndexedBook` ships them straight through to the writer.
-pub fn parse_ebook_targets(targets: Vec<ParseTarget>, opts: ScanOptions) -> Vec<IndexedBook> {
-    targets
-        .into_iter()
-        .map(|t| {
-            let mut book = extract_metadata(&t.absolute, t.filename, &opts);
-            book.mtime_epoch = t.mtime_epoch;
-            book.size_bytes = t.size_bytes;
-            book
-        })
-        .collect()
-}
-
-/// `entry.metadata()` + `modified()` → epoch seconds. Anything we can't
-/// stat returns `(0, 0)`, which the diff treats as the "Backfill" sentinel
-/// — same as a freshly migrated row. That's fine: if we couldn't stat it
-/// now, the next run will either succeed (and the row updates) or keep
-/// failing (and the row stays in the same Backfill-ish state). No data loss.
-fn stat_file(path: &Path) -> (i64, i64) {
-    let Ok(meta) = std::fs::metadata(path) else {
-        return (0, 0);
-    };
-    let size = meta.len() as i64;
-    let mtime = meta
-        .modified()
-        .ok()
-        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-        .map(|d| d.as_secs() as i64)
-        .unwrap_or(0);
-    (mtime, size)
-}
-
-fn extract_metadata(path: &Path, filename: String, opts: &ScanOptions) -> IndexedBook {
-    let mut doc = match EpubDoc::new(path) {
-        Ok(d) => d,
-        Err(e) => {
-            return IndexedBook {
-                metadata: EbookMetadata {
-                    filename,
-                    error: Some(format!("could not open epub: {e}")),
-                    ..Default::default()
-                },
-                cover: None,
-                // Stat values get overwritten by `parse_ebook_targets`
-                // before the writer sees this struct.
-                mtime_epoch: 0,
-                size_bytes: 0,
-            };
-        }
-    };
-
-    let creators = collect_contributors(&doc, "creator");
-    let contributors = collect_contributors(&doc, "contributor");
-    let identifiers = collect_identifiers(&doc);
-    let (series, series_index) = collect_series(&doc);
-
-    let cover = resolve_cover(path, &mut doc, opts);
-    let accent = cover
-        .as_ref()
-        .and_then(|(_mime, bytes)| extract_accent(bytes));
-
-    IndexedBook {
-        metadata: EbookMetadata {
-            id: 0,
-            filename,
-            title: first(&doc, "title"),
-            description: first(&doc, "description"),
-            publisher: first(&doc, "publisher"),
-            published: first(&doc, "date"),
-            modified: first(&doc, "dcterms:modified"),
-            language: first(&doc, "language"),
-            rights: first(&doc, "rights"),
-            source: first(&doc, "source"),
-            coverage: first(&doc, "coverage"),
-            dc_type: first(&doc, "type"),
-            dc_format: first(&doc, "format"),
-            relation: first(&doc, "relation"),
-
-            creators,
-            contributors,
-            subjects: all(&doc, "subject"),
-            identifiers,
-
-            series,
-            series_index,
-            series_id: None,
-
-            epub_version: Some(format_version(doc.version)),
-            unique_identifier: doc.unique_identifier.clone(),
-            resource_count: doc.resources.len(),
-            spine_count: doc.spine.len(),
-            toc_count: doc.toc.len(),
-
-            cover_url: None,
-            accent,
-            formats: vec![],
-            added_at: None,
-            error: None,
-            has_override: false,
-        },
-        cover,
-        // Stat values get overwritten by `parse_ebook_targets` before the
-        // writer sees this struct.
-        mtime_epoch: 0,
-        size_bytes: 0,
-    }
-}
-
-/// Sidecar-first cover resolution.
-///
-/// 1. If `<path>` has a sidecar (per-stem first, folder-level fallback),
-///    read its bytes and return them.
-/// 2. Otherwise, ask the EPUB for its embedded cover.
-/// 3. If `opts.materialize_sidecars` is set and the embedded cover came back
-///    successfully, write it to a `<basename>.{jpg|png}` sidecar so the next
-///    scan hits the sidecar directly. Failures are non-fatal.
-///
-/// Returns the cover bytes used for *this* scan (the in-memory copy, even
-/// when materialization wrote them to disk — this avoids a round-trip read).
-fn resolve_cover<R: std::io::Read + std::io::Seek>(
-    path: &Path,
-    doc: &mut EpubDoc<R>,
-    opts: &ScanOptions,
-) -> Option<(String, Vec<u8>)> {
-    let mut corrupt_sidecar: Option<PathBuf> = None;
-    if let Some(sidecar) = library_layout::sidecar_cover_for(path) {
-        if let Some(bytes) = read_sidecar(&sidecar) {
-            return Some(bytes);
-        }
-        // Sidecar lookup found a file but reading it failed — fall through
-        // to the embedded path. Pass the broken path to materialize_sidecar
-        // so it can repair the cache instead of refusing forever.
-        corrupt_sidecar = Some(sidecar);
-    }
-
-    let embedded = doc.get_cover().map(|(bytes, mime)| {
-        let mime = if mime.is_empty() {
-            "image/jpeg".to_string()
-        } else {
-            mime
-        };
-        (mime, bytes)
-    });
-
-    if opts.materialize_sidecars {
-        if let Some((mime, bytes)) = embedded.as_ref() {
-            materialize_sidecar(path, mime, bytes, corrupt_sidecar.as_deref());
-        }
-    }
-
-    embedded
-}
-
-fn read_sidecar(path: &Path) -> Option<(String, Vec<u8>)> {
-    let bytes = std::fs::read(path).ok()?;
-    // A zero-length file is no better than a read error — surfacing it as
-    // "the cover" would just blank the row in the UI. Treat as corrupt so
-    // the materialize path can repair it next pass.
-    if bytes.is_empty() {
-        return None;
-    }
-    let mime = mime_for_extension(path).to_string();
-    Some((mime, bytes))
-}
-
-fn mime_for_extension(path: &Path) -> &'static str {
-    match path
-        .extension()
-        .and_then(|e| e.to_str())
-        .map(|e| e.to_ascii_lowercase())
-        .as_deref()
-    {
-        Some("png") => "image/png",
-        // jpg / jpeg / anything else falls back to JPEG. Embedded EPUB covers
-        // are overwhelmingly JPEG, and a wrong-but-close mime is better than
-        // none for the cover endpoint.
-        _ => "image/jpeg",
-    }
-}
-
-/// Best-effort write of `<basename>.{jpg|png}` next to the epub so future
-/// scans skip the zip. Fails silently — this is a cache, not a contract.
-///
-/// `corrupt_sidecar`, when set, is a sidecar path the caller already
-/// confirmed is unreadable. We allow overwriting *exactly* that file so a
-/// corrupt cache entry self-heals on the next scan instead of forcing every
-/// future scan to re-open the zip. Anything else under `target.exists()`
-/// (a valid file, a different filename) we leave alone.
-fn materialize_sidecar(epub_path: &Path, mime: &str, bytes: &[u8], corrupt_sidecar: Option<&Path>) {
-    let Some(parent) = epub_path.parent() else {
-        return;
-    };
-    let Some(stem) = epub_path.file_stem().and_then(|s| s.to_str()) else {
-        return;
-    };
-    let ext = if mime.eq_ignore_ascii_case("image/png") {
-        "png"
-    } else {
-        "jpg"
-    };
-    let target = parent.join(format!("{stem}.{ext}"));
-    if target.exists() {
-        let is_known_corrupt = corrupt_sidecar.is_some_and(|p| p == target.as_path());
-        if !is_known_corrupt {
-            // A valid file we don't own (race or user-dropped sidecar). Don't
-            // clobber.
-            return;
-        }
-        // Fall through and overwrite — std::fs::write truncates the existing
-        // file, repairing the cache.
-        tracing::warn!(
-            path = %target.display(),
-            "repairing unreadable cover sidecar"
-        );
-    }
-    if let Err(e) = std::fs::write(&target, bytes) {
-        tracing::warn!(
-            error = %e,
-            path = %target.display(),
-            "could not materialize cover sidecar; falling back to embedded for this scan"
-        );
-    }
-}
-
-fn format_version(v: EpubVersion) -> String {
-    match v {
-        EpubVersion::Version2_0 => "2.0".to_string(),
-        EpubVersion::Version3_0 => "3.0".to_string(),
-        other => format!("{other:?}"),
-    }
-}
-
-fn collect_contributors<R: std::io::Read + std::io::Seek>(
-    doc: &EpubDoc<R>,
-    key: &str,
-) -> Vec<Contributor> {
-    doc.metadata
-        .iter()
-        .filter(|m| m.property == key)
-        .filter_map(|m| {
-            let name = m.value.trim().to_string();
-            if name.is_empty() {
-                return None;
-            }
-            let role = m
-                .refinement("role")
-                .map(|r| r.value.clone())
-                .or_else(|| lookup_refinement(&m.refined, "role"));
-            let file_as = m
-                .refinement("file-as")
-                .map(|r| r.value.clone())
-                .or_else(|| lookup_refinement(&m.refined, "file-as"));
-            Some(Contributor {
-                name,
-                role,
-                file_as,
-                id: None,
-            })
-        })
-        .collect()
-}
-
-fn lookup_refinement(refs: &[epub::doc::MetadataRefinement], key: &str) -> Option<String> {
-    refs.iter()
-        .find(|r| r.property == key)
-        .map(|r| r.value.clone())
-}
-
-/// Resolve (series, series_index) from the OPF.
-///
-/// EPUB3 stores a series as a `belongs-to-collection` metadata entry whose
-/// `group-position` refinement holds the index. Calibre's legacy EPUB2 tooling
-/// writes top-level `<meta name="calibre:series">` and `calibre:series_index`
-/// entries instead. We try EPUB3 first (with the refinement), then fall back
-/// to the Calibre keys.
-fn collect_series<R: std::io::Read + std::io::Seek>(
-    doc: &EpubDoc<R>,
-) -> (Option<String>, Option<String>) {
-    if let Some(m) = doc
-        .metadata
-        .iter()
-        .find(|m| m.property == "belongs-to-collection")
-    {
-        let name = m.value.trim().to_string();
-        if !name.is_empty() {
-            let idx = lookup_refinement(&m.refined, "group-position")
-                .map(|s| s.trim().to_string())
-                .filter(|s| !s.is_empty());
-            return (Some(name), idx);
-        }
-    }
-    (
-        first(doc, "calibre:series"),
-        first(doc, "calibre:series_index"),
-    )
-}
-
-fn collect_identifiers<R: std::io::Read + std::io::Seek>(doc: &EpubDoc<R>) -> Vec<Identifier> {
-    doc.metadata
-        .iter()
-        .filter(|m| m.property == "identifier")
-        .filter_map(|m| {
-            let value = m.value.trim().to_string();
-            if value.is_empty() {
-                return None;
-            }
-            let scheme = lookup_refinement(&m.refined, "scheme")
-                .or_else(|| lookup_refinement(&m.refined, "identifier-type"));
-            Some(Identifier { value, scheme })
-        })
-        .collect()
-}
-
-fn first<R: std::io::Read + std::io::Seek>(doc: &EpubDoc<R>, key: &str) -> Option<String> {
-    doc.metadata
-        .iter()
-        .find(|m| m.property == key)
-        .map(|m| m.value.trim().to_string())
-        .filter(|s| !s.is_empty())
-}
-
-fn all<R: std::io::Read + std::io::Seek>(doc: &EpubDoc<R>, key: &str) -> Vec<String> {
-    doc.metadata
-        .iter()
-        .filter(|m| m.property == key)
-        .map(|m| m.value.trim().to_string())
-        .filter(|s| !s.is_empty())
-        .collect()
-}
-
-/// Hard cap on embedded cover bytes we'll hand to the `image` decoder.
-/// Higher than the 10 MiB HTTP cap in `author_photos` because these are
-/// trusted local files; 20 MiB still covers uncompressed print-resolution
-/// covers while bounding the decode allocation against a crafted EPUB.
-const MAX_EMBEDDED_COVER_BYTES: usize = 20 * 1024 * 1024;
-
-/// F1.7 Atrium: extract a representative accent color from cover bytes.
-/// Returns an `oklch(L C H)` string clamped to a readable band, or `None`
-/// when decoding fails or the cover has no chromatic content. See the
-/// [Atrium design doc](../../../docs/design/atrium-design-system.md) §2b
-/// for the algorithm rationale (hue-bucket → highest-weighted → OKLCH).
-pub fn extract_accent(bytes: &[u8]) -> Option<String> {
-    if bytes.is_empty() {
-        return None;
-    }
-    if bytes.len() > MAX_EMBEDDED_COVER_BYTES {
-        return None;
-    }
-    let img = image::load_from_memory(bytes).ok()?;
-    let small = img.thumbnail(32, 48).to_rgb8();
-
-    let mut buckets = [AccentBucket::default(); 12];
-    for px in small.pixels() {
-        let r = px[0] as f32 / 255.0;
-        let g = px[1] as f32 / 255.0;
-        let b = px[2] as f32 / 255.0;
-        let max = r.max(g).max(b);
-        let min = r.min(g).min(b);
-        let delta = max - min;
-        // Filter near-black and near-grayscale pixels so they don't
-        // overwhelm the actual artwork color.
-        if max < 0.10 || delta < 0.06 {
-            continue;
-        }
-        let hue = if delta == 0.0 {
-            0.0
-        } else if (max - r).abs() < f32::EPSILON {
-            60.0 * (((g - b) / delta).rem_euclid(6.0))
-        } else if (max - g).abs() < f32::EPSILON {
-            60.0 * ((b - r) / delta + 2.0)
-        } else {
-            60.0 * ((r - g) / delta + 4.0)
-        };
-        let sat = if max == 0.0 { 0.0 } else { delta / max };
-        let weight = sat * max;
-        let idx = ((hue / 30.0).floor() as usize).min(11);
-        buckets[idx].r += r * weight;
-        buckets[idx].g += g * weight;
-        buckets[idx].b += b * weight;
-        buckets[idx].w += weight;
-    }
-
-    let best = buckets
-        .iter()
-        .max_by(|a, b| a.w.partial_cmp(&b.w).unwrap_or(std::cmp::Ordering::Equal))?;
-    if best.w == 0.0 {
-        return None;
-    }
-    let r = best.r / best.w;
-    let g = best.g / best.w;
-    let b = best.b / best.w;
-    let (l, c, h) = rgb_to_oklch(r, g, b);
-    let l = l.clamp(0.55, 0.78);
-    let c = c.clamp(0.06, 0.18);
-    Some(format!("oklch({l:.3} {c:.3} {h:.1})"))
-}
-
-#[derive(Default, Clone, Copy)]
-struct AccentBucket {
-    r: f32,
-    g: f32,
-    b: f32,
-    w: f32,
-}
-
-/// Convert non-linear sRGB in [0, 1] to OKLCH. Matrix from Björn Ottosson,
-/// <https://bottosson.github.io/posts/oklab/>. Returns `(L, C, H°)`.
-fn rgb_to_oklch(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
-    fn linearize(v: f32) -> f32 {
-        if v <= 0.04045 {
-            v / 12.92
-        } else {
-            ((v + 0.055) / 1.055).powf(2.4)
-        }
-    }
-    // OKLab matrix coefficients (Björn Ottosson). Truncated to f32 precision
-    // — clippy's `excessive_precision` lint flags the full 10-digit form,
-    // and the extra digits don't survive `f32` round-off anyway.
-    let r = linearize(r);
-    let g = linearize(g);
-    let b = linearize(b);
-    let l = 0.412_221_47 * r + 0.536_332_54 * g + 0.051_445_995 * b;
-    let m = 0.211_903_5 * r + 0.680_699_5 * g + 0.107_396_96 * b;
-    let s = 0.088_302_46 * r + 0.281_718_85 * g + 0.629_978_7 * b;
-    let l_ = l.cbrt();
-    let m_ = m.cbrt();
-    let s_ = s.cbrt();
-    let big_l = 0.210_454_26 * l_ + 0.793_617_8 * m_ - 0.004_072_047 * s_;
-    let a = 1.977_998_5 * l_ - 2.428_592_2 * m_ + 0.450_593_7 * s_;
-    let b_oklab = 0.025_904_037 * l_ + 0.782_771_77 * m_ - 0.808_675_77 * s_;
-    let c = (a * a + b_oklab * b_oklab).sqrt();
-    let mut h = b_oklab.atan2(a).to_degrees();
-    if h < 0.0 {
-        h += 360.0;
-    }
-    (big_l, c, h)
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
+pub(crate) mod test_support {
+    use std::path::Path;
     use std::sync::atomic::{AtomicUsize, Ordering};
 
     /// Build a unique temp directory per test invocation. Rust runs unit
     /// tests in parallel by default, so a fixed path under `temp_dir()`
     /// would collide between tests (and between repeated runs).
-    fn make_test_dir(suffix: &str) -> std::path::PathBuf {
+    pub(crate) fn make_test_dir(suffix: &str) -> std::path::PathBuf {
         static COUNTER: AtomicUsize = AtomicUsize::new(0);
         let pid = std::process::id();
         let seq = COUNTER.fetch_add(1, Ordering::Relaxed);
@@ -746,18 +155,9 @@ mod tests {
         dir
     }
 
-    #[test]
-    fn scan_with_no_path_returns_empty() {
-        let out = scan_ebook_library(None);
-        assert!(out.books.is_empty());
-        assert!(out.path.is_none());
-    }
-
-    // ── F1.7 accent extraction ─────────────────────────────────────────
-
     /// Build an in-memory PNG of a single solid color. Used to drive
     /// `extract_accent` without baking a fixture cover into the repo.
-    fn solid_color_png(r: u8, g: u8, b: u8, w: u32, h: u32) -> Vec<u8> {
+    pub(crate) fn solid_color_png(r: u8, g: u8, b: u8, w: u32, h: u32) -> Vec<u8> {
         let img = image::RgbImage::from_pixel(w, h, image::Rgb([r, g, b]));
         let mut bytes = Vec::new();
         image::DynamicImage::ImageRgb8(img)
@@ -769,216 +169,10 @@ mod tests {
         bytes
     }
 
-    #[test]
-    fn extract_accent_returns_oklch_for_saturated_cover() {
-        let bytes = solid_color_png(200, 60, 50, 64, 96);
-        let accent = extract_accent(&bytes).expect("saturated cover yields accent");
-        assert!(
-            accent.starts_with("oklch("),
-            "accent should be an oklch() string, got {accent}"
-        );
-        // Clamps in extract_accent require the L value to stay readable.
-        let mid = accent
-            .trim_start_matches("oklch(")
-            .trim_end_matches(')')
-            .split_whitespace()
-            .next()
-            .unwrap()
-            .parse::<f32>()
-            .unwrap();
-        assert!(
-            (0.55..=0.78).contains(&mid),
-            "lightness {mid} should be clamped to [0.55, 0.78]"
-        );
-    }
-
-    #[test]
-    fn extract_accent_returns_none_for_empty_bytes() {
-        assert!(extract_accent(&[]).is_none());
-    }
-
-    #[test]
-    fn extract_accent_returns_none_for_corrupt_bytes() {
-        assert!(extract_accent(b"not an image, just text").is_none());
-    }
-
-    #[test]
-    fn extract_accent_returns_none_for_oversized_bytes() {
-        // Start from a *valid* cover that decodes to an accent on its own, so
-        // the only thing keeping the oversized case out of the decoder is the
-        // size guard. (All-zero bytes would fail to decode regardless, which
-        // wouldn't catch a regression that removed the guard.)
-        let mut bytes = solid_color_png(200, 60, 50, 64, 96);
-        assert!(
-            extract_accent(&bytes).is_some(),
-            "sanity: the unpadded cover should decode to an accent"
-        );
-        // Pad past the cap with trailing bytes the PNG decoder ignores (it
-        // stops at IEND). Without the guard these would still decode and
-        // yield Some, so this assertion fails if the cap check is removed.
-        bytes.resize(MAX_EMBEDDED_COVER_BYTES + 1, 0);
-        assert!(
-            extract_accent(&bytes).is_none(),
-            "oversized cover bytes should be rejected by the size guard before decoding"
-        );
-    }
-
-    #[test]
-    fn extract_accent_returns_none_for_pure_black() {
-        let bytes = solid_color_png(0, 0, 0, 64, 96);
-        assert!(
-            extract_accent(&bytes).is_none(),
-            "all-black cover should produce no accent"
-        );
-    }
-
-    #[test]
-    fn extract_accent_returns_none_for_pure_gray() {
-        let bytes = solid_color_png(128, 128, 128, 64, 96);
-        assert!(
-            extract_accent(&bytes).is_none(),
-            "grayscale cover should produce no accent (no chroma)"
-        );
-    }
-
-    #[test]
-    fn extract_accent_completes_within_budget() {
-        // Real EPUB covers top out around 1500×2250. Test against that size
-        // with a generous debug-mode budget — release builds run ~3× faster.
-        // The point of the test is to catch a regression to seconds, not to
-        // hold a tight production-grade budget in unoptimized builds. GitHub
-        // Actions ubuntu-latest runners have measurably slower per-core
-        // throughput than developer workstations and can spend 600–700 ms on
-        // this input in debug builds, so the threshold is set to 2 s — any
-        // multi-second regression still trips it.
-        let bytes = solid_color_png(80, 140, 200, 1500, 2250);
-        let start = std::time::Instant::now();
-        let _ = extract_accent(&bytes);
-        let elapsed = start.elapsed();
-        assert!(
-            elapsed < std::time::Duration::from_secs(2),
-            "extract_accent must stay well under multiple seconds on realistic input; took {elapsed:?}"
-        );
-    }
-
-    #[test]
-    fn scan_with_missing_path_reports_error() {
-        let out = scan_ebook_library(Some("/definitely/does/not/exist/omnibus_ebook_test"));
-        assert!(out.error.is_some());
-    }
-
-    #[test]
-    fn scan_ignores_non_epub_files() {
-        let dir = make_test_dir("ignore");
-        std::fs::write(dir.join("notes.txt"), b"hi").unwrap();
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        std::fs::remove_dir_all(&dir).unwrap();
-        assert!(out.books.is_empty());
-        assert!(out.error.is_none());
-    }
-
-    #[test]
-    fn scan_recurses_into_subdirectories() {
-        let dir = make_test_dir("recursive");
-        let nested = dir.join("series").join("vol1");
-        std::fs::create_dir_all(&nested).unwrap();
-        std::fs::write(dir.join("top.epub"), b"not a zip").unwrap();
-        std::fs::write(nested.join("deep.epub"), b"not a zip").unwrap();
-        std::fs::write(nested.join("cover.jpg"), b"ignore").unwrap();
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        std::fs::remove_dir_all(&dir).unwrap();
-        // Two epubs + zero synthetic dir errors under a healthy tree.
-        let book_entries: Vec<&str> = out
-            .books
-            .iter()
-            .filter(|b| b.metadata.error.is_none() || !b.metadata.filename.is_empty())
-            .map(|b| b.metadata.filename.as_str())
-            .collect();
-        assert_eq!(out.books.len(), 2);
-        assert!(book_entries.contains(&"top.epub"));
-        assert!(book_entries
-            .iter()
-            .any(|n| n.ends_with("deep.epub") && n.contains("series")));
-    }
-
-    #[test]
-    fn scan_records_parse_errors_per_file() {
-        let dir = make_test_dir("bad");
-        std::fs::write(dir.join("broken.epub"), b"not actually a zip").unwrap();
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        std::fs::remove_dir_all(&dir).unwrap();
-        assert_eq!(out.books.len(), 1);
-        assert!(out.books[0].metadata.error.is_some());
-        assert_eq!(out.books[0].metadata.filename, "broken.epub");
-    }
-
-    // Permission tests only run where we can reliably strip read access
-    // from a directory — on Windows, even removing all permission bits
-    // doesn't reliably make `read_dir` fail as the owning process.
-    #[cfg(unix)]
-    #[test]
-    fn scan_continues_past_unreadable_subdirectory() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = make_test_dir("unreadable_subdir");
-        std::fs::write(dir.join("good.epub"), b"not a zip").unwrap();
-        let locked = dir.join("locked");
-        std::fs::create_dir_all(&locked).unwrap();
-        // Drop all permissions so read_dir on the subdir fails. Skip the
-        // test on platforms where this doesn't take effect (e.g. running
-        // as root in CI containers) rather than asserting a false
-        // positive.
-        std::fs::set_permissions(&locked, std::fs::Permissions::from_mode(0o000)).unwrap();
-        if std::fs::read_dir(&locked).is_ok() {
-            std::fs::set_permissions(&locked, std::fs::Permissions::from_mode(0o755)).unwrap();
-            std::fs::remove_dir_all(&dir).unwrap();
-            return;
-        }
-
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-
-        // Restore permissions before removing the tree, otherwise cleanup
-        // fails on some platforms.
-        std::fs::set_permissions(&locked, std::fs::Permissions::from_mode(0o755)).unwrap();
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        // Top-level scan must not report a fatal error.
-        assert!(out.error.is_none());
-        // Good epub still surfaces.
-        assert!(out.books.iter().any(|b| b.metadata.filename == "good.epub"));
-        // Locked subdir surfaces as a synthetic error entry, not silently
-        // dropped, and exactly once — Phase B must skip the empty-uuid
-        // placeholder so it doesn't get parsed as a fake epub AND appended
-        // as an error row (which would produce two rows for the same dir).
-        let locked_rows: Vec<_> = out
-            .books
-            .iter()
-            .filter(|b| b.metadata.filename == "locked")
-            .collect();
-        assert_eq!(
-            locked_rows.len(),
-            1,
-            "exactly one error row per unreadable subdir, got {locked_rows:?}",
-        );
-        let msg = locked_rows[0].metadata.error.as_deref().unwrap_or("");
-        assert!(
-            msg.contains("could not read directory"),
-            "error message should keep the legacy prefix, got {msg:?}",
-        );
-        // Underlying io::Error detail (e.g. "permission denied") must be
-        // carried through so callers can distinguish failure modes.
-        assert!(
-            msg.len() > "could not read directory".len(),
-            "error message should include the io detail, got {msg:?}",
-        );
-    }
-
-    // ---------- Sidecar cover (F0.6) ----------
-
     /// Path to a real fixture epub from `test_data/epubs/generated/` so
     /// cover-related tests have real OPF + embedded image bytes to work
     /// with. Stub `b"not a zip"` files won't decode.
-    fn fixture(name: &str) -> std::path::PathBuf {
+    pub(crate) fn fixture(name: &str) -> std::path::PathBuf {
         std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
             .join("..")
             .join("test_data")
@@ -988,64 +182,16 @@ mod tests {
     }
 
     /// Copy a fixture epub into `dest` and return the destination path.
-    fn copy_fixture_into(name: &str, dest: &Path) -> std::path::PathBuf {
+    pub(crate) fn copy_fixture_into(name: &str, dest: &Path) -> std::path::PathBuf {
         let target = dest.join(name);
         std::fs::copy(fixture(name), &target).expect("copy fixture");
         target
     }
 
-    #[test]
-    fn extract_metadata_uses_sidecar_when_present() {
-        // alpha.epub ships an embedded cover. Plant a recognizably-different
-        // sidecar next to it; the scanner must return the sidecar bytes.
-        let dir = make_test_dir("sidecar_wins");
-        copy_fixture_into("alpha.epub", &dir);
-        let sidecar_bytes: &[u8] = b"sidecar-jpg-magic-bytes";
-        std::fs::write(dir.join("alpha.jpg"), sidecar_bytes).unwrap();
-
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        let alpha = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "alpha.epub")
-            .expect("alpha present");
-        let (mime, bytes) = alpha.cover.as_ref().expect("cover present");
-        assert_eq!(bytes, sidecar_bytes, "expected sidecar bytes, got embedded");
-        assert_eq!(mime, "image/jpeg");
-    }
-
-    #[test]
-    fn extract_metadata_uses_embedded_when_no_sidecar() {
-        // alpha.epub has an embedded cover; no sidecar planted. We don't
-        // know the exact embedded bytes, but they should be non-empty and
-        // the cover slot must be populated. Default ScanOptions disables
-        // materialization, so no sidecar should appear after the scan.
-        let dir = make_test_dir("embedded_only");
-        copy_fixture_into("alpha.epub", &dir);
-
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        let sidecar_appeared = find_materialized_sidecar(&dir, "alpha").is_some();
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert!(
-            !sidecar_appeared,
-            "default ScanOptions must not materialize sidecars"
-        );
-        let alpha = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "alpha.epub")
-            .expect("alpha present");
-        let (_, bytes) = alpha.cover.as_ref().expect("embedded cover present");
-        assert!(!bytes.is_empty());
-    }
-
     /// Locate the materialized sidecar in `dir` for `<stem>` — checks both
     /// `.jpg` and `.png` since the materialized extension follows the
     /// embedded mime, which the test can't predict for arbitrary fixtures.
-    fn find_materialized_sidecar(dir: &Path, stem: &str) -> Option<std::path::PathBuf> {
+    pub(crate) fn find_materialized_sidecar(dir: &Path, stem: &str) -> Option<std::path::PathBuf> {
         for ext in ["jpg", "jpeg", "png"] {
             let candidate = dir.join(format!("{stem}.{ext}"));
             if candidate.exists() {
@@ -1053,272 +199,5 @@ mod tests {
             }
         }
         None
-    }
-
-    #[test]
-    fn extract_metadata_materializes_sidecar_with_opt_in() {
-        // With `materialize_sidecars: true`, scanning an epub that has an
-        // embedded cover but no sidecar must write `<basename>.{jpg|png}`
-        // (extension matches embedded mime) next to the file so subsequent
-        // scans hit the sidecar directly.
-        let dir = make_test_dir("materialize");
-        copy_fixture_into("alpha.epub", &dir);
-        assert!(
-            find_materialized_sidecar(&dir, "alpha").is_none(),
-            "precondition: no sidecar yet"
-        );
-
-        let out = scan_ebook_library_with(
-            Some(dir.to_str().unwrap()),
-            ScanOptions {
-                materialize_sidecars: true,
-            },
-        );
-
-        let sidecar = find_materialized_sidecar(&dir, "alpha");
-        let written = sidecar.as_ref().and_then(|p| std::fs::read(p).ok());
-        let alpha = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "alpha.epub")
-            .map(|b| b.cover.as_ref().map(|(_, bytes)| bytes.clone()))
-            .unwrap_or(None);
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert!(sidecar.is_some(), "sidecar should have been written");
-        assert_eq!(
-            written.as_deref(),
-            alpha.as_deref(),
-            "written sidecar bytes must match returned cover bytes"
-        );
-    }
-
-    #[test]
-    fn extract_metadata_second_scan_reads_sidecar_not_zip() {
-        // After materialization, swap the sidecar with different bytes. The
-        // next scan should return *those* bytes, proving the read came from
-        // the sidecar and not the unchanged embedded cover in the zip.
-        let dir = make_test_dir("second_scan");
-        copy_fixture_into("alpha.epub", &dir);
-
-        // First scan: materialize.
-        let _ = scan_ebook_library_with(
-            Some(dir.to_str().unwrap()),
-            ScanOptions {
-                materialize_sidecars: true,
-            },
-        );
-
-        let sidecar_path =
-            find_materialized_sidecar(&dir, "alpha").expect("first scan materialized a sidecar");
-
-        // Replace the sidecar (same path/extension) with sentinel bytes.
-        let sentinel: &[u8] = b"replaced-after-materialization";
-        std::fs::write(&sidecar_path, sentinel).unwrap();
-
-        // Second scan (default opts) — should read the sentinel, not
-        // re-extract the embedded cover.
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        let alpha = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "alpha.epub")
-            .expect("alpha present");
-        let (_, bytes) = alpha.cover.as_ref().expect("cover present");
-        assert_eq!(
-            bytes, sentinel,
-            "second scan should have read the swapped sidecar"
-        );
-    }
-
-    #[test]
-    fn extract_metadata_repairs_unreadable_sidecar_on_materialize() {
-        // alpha.epub has an embedded cover. Plant a zero-length sidecar that
-        // sidecar_cover_for() will pick up but read_sidecar() can't use.
-        // With materialize_sidecars=true, the broken cache must be repaired
-        // so the next scan reads the sidecar instead of re-opening the zip.
-        let dir = make_test_dir("repair_sidecar");
-        copy_fixture_into("alpha.epub", &dir);
-
-        // alpha.epub embeds a PNG, so the materializer would write
-        // `alpha.png`. Plant the corrupt sidecar at that exact path.
-        let broken = dir.join("alpha.png");
-        std::fs::write(&broken, b"").unwrap();
-
-        let out = scan_ebook_library_with(
-            Some(dir.to_str().unwrap()),
-            ScanOptions {
-                materialize_sidecars: true,
-            },
-        );
-
-        let repaired_bytes = std::fs::read(&broken).expect("sidecar still on disk");
-        let alpha_cover = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "alpha.epub")
-            .and_then(|b| b.cover.as_ref().map(|(_, bytes)| bytes.clone()));
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert!(
-            !repaired_bytes.is_empty(),
-            "broken zero-length sidecar should have been repaired"
-        );
-        assert_eq!(
-            alpha_cover.as_deref(),
-            Some(repaired_bytes.as_slice()),
-            "repaired sidecar bytes must match the embedded cover the scan returned"
-        );
-    }
-
-    #[test]
-    fn extract_metadata_does_not_clobber_unrelated_existing_sidecar() {
-        // The repair gate must only overwrite the *exact* corrupt file the
-        // sidecar lookup returned — never a different valid file that
-        // happens to sit at the materialize target.
-        //
-        // Setup: alpha.epub embeds a PNG, so materialize would target
-        // alpha.png. We plant a corrupt (empty) `alpha.jpg` (which jpg-over-
-        // png priority makes sidecar_cover_for return) AND a valid `alpha.png`
-        // (the user's curated cover). The materialize step must refuse to
-        // overwrite alpha.png because the *known* corrupt path is alpha.jpg.
-        let dir = make_test_dir("no_clobber");
-        copy_fixture_into("alpha.epub", &dir);
-
-        let corrupt_jpg = dir.join("alpha.jpg");
-        std::fs::write(&corrupt_jpg, b"").unwrap();
-        let valid_png = dir.join("alpha.png");
-        let curated: &[u8] = b"user-curated-cover-do-not-touch";
-        std::fs::write(&valid_png, curated).unwrap();
-
-        let _ = scan_ebook_library_with(
-            Some(dir.to_str().unwrap()),
-            ScanOptions {
-                materialize_sidecars: true,
-            },
-        );
-
-        let png_after = std::fs::read(&valid_png).unwrap();
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert_eq!(
-            png_after, curated,
-            "alpha.png is not the corrupt sidecar — must not be overwritten"
-        );
-    }
-
-    #[test]
-    fn extract_metadata_no_embedded_no_sidecar_returns_none() {
-        // gamma.epub has no embedded cover. No sidecar planted, no
-        // materialization. Cover should stay None and no file should be
-        // written.
-        let dir = make_test_dir("no_cover");
-        copy_fixture_into("gamma.epub", &dir);
-
-        let out = scan_ebook_library_with(
-            Some(dir.to_str().unwrap()),
-            ScanOptions {
-                materialize_sidecars: true,
-            },
-        );
-        let sidecar_appeared = find_materialized_sidecar(&dir, "gamma").is_some();
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert!(
-            !sidecar_appeared,
-            "no embedded cover → nothing to materialize"
-        );
-        let gamma = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "gamma.epub")
-            .expect("gamma present");
-        assert!(gamma.cover.is_none());
-    }
-
-    #[test]
-    fn scan_handles_calibre_shaped_tree_and_ignores_metadata_opf() {
-        // Lock in the read-tolerance promise from F0.6: a Calibre-style
-        // library tree (`<Lastname, First>/Title (id)/title.epub` plus an
-        // adjacent `metadata.opf` Calibre wrote out) must scan correctly.
-        // We assert (a) the epub is found, (b) the title comes from the
-        // *embedded* OPF inside the epub, not the deliberately-wrong
-        // sidecar `metadata.opf`. The sidecar is ignored entirely.
-        let dir = make_test_dir("calibre_shaped");
-        let book_dir = dir.join("Lovelace, Ada").join("Alpha (42)");
-        std::fs::create_dir_all(&book_dir).unwrap();
-        std::fs::copy(fixture("alpha.epub"), book_dir.join("alpha.epub")).unwrap();
-        // Calibre's metadata.opf — write garbage into it so any code path
-        // that *did* read it would visibly disagree with the embedded OPF.
-        std::fs::write(
-            book_dir.join("metadata.opf"),
-            br#"<?xml version="1.0"?>
-<package xmlns="http://www.idpf.org/2007/opf" version="2.0">
-<metadata xmlns:dc="http://purl.org/dc/elements/1.1/">
-<dc:title>WRONG TITLE FROM CALIBRE SIDECAR</dc:title>
-<dc:creator>Wrong Author</dc:creator>
-</metadata></package>"#,
-        )
-        .unwrap();
-
-        let out = scan_ebook_library(Some(dir.to_str().unwrap()));
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert!(out.error.is_none(), "scan errored: {:?}", out.error);
-        // Exactly one epub found despite the metadata.opf sibling.
-        let epubs: Vec<_> = out
-            .books
-            .iter()
-            .filter(|b| b.metadata.error.is_none())
-            .collect();
-        assert_eq!(epubs.len(), 1);
-        // Title comes from the embedded OPF, not the misleading sidecar.
-        assert_eq!(epubs[0].metadata.title.as_deref(), Some("Alpha"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn extract_metadata_materialization_failure_falls_back_to_embedded() {
-        // chmod the directory read-only-execute so write fails. The scanner
-        // must still return cover bytes (from embedded), and no sidecar
-        // should appear.
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = make_test_dir("readonly_dir");
-        copy_fixture_into("alpha.epub", &dir);
-        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(0o555)).unwrap();
-
-        // Skip if the chmod didn't take (e.g. running as root in some CI
-        // containers).
-        if std::fs::write(dir.join("write_probe"), b"x").is_ok() {
-            std::fs::remove_file(dir.join("write_probe")).ok();
-            std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(0o755)).unwrap();
-            std::fs::remove_dir_all(&dir).unwrap();
-            return;
-        }
-
-        let out = scan_ebook_library_with(
-            Some(dir.to_str().unwrap()),
-            ScanOptions {
-                materialize_sidecars: true,
-            },
-        );
-
-        let sidecar_appeared = find_materialized_sidecar(&dir, "alpha").is_some();
-
-        // Restore perms before cleanup.
-        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(0o755)).unwrap();
-        std::fs::remove_dir_all(&dir).unwrap();
-
-        assert!(!sidecar_appeared, "read-only fs must not produce a sidecar");
-        let alpha = out
-            .books
-            .iter()
-            .find(|b| b.metadata.filename == "alpha.epub")
-            .expect("alpha present");
-        let (_, bytes) = alpha.cover.as_ref().expect("embedded fallback present");
-        assert!(!bytes.is_empty(), "embedded fallback must be non-empty");
     }
 }
